lib: updated well-known-types protos

+++ lib/src/commonMain/kotlin/com/google/protobuf/duration.builder.kt
@@ -4,16 +4,16 @@ import dev.mpr.grpc.ProtoDsl
 
 @ProtoDsl
 class DurationBuilder {
-	constructor()
+    constructor()
 
-	constructor(copy: Duration) {
+    constructor(copy: Duration) {
 	    builderCopy.seconds = copy.seconds
 	    builderCopy.nanos = copy.nanos
 	    builderCopy.unknownFields = copy.unknownFields
 	}
 
 	private object builderCopy {
-	    var seconds: Int = 0
+	    var seconds: Long = 0L
 	    var nanos: Int = 0
 	    var unknownFields: ByteArray? = null
 	}
@@ -24,20 +24,21 @@ class DurationBuilder {
     	builderCopy.unknownFields
 	)
 	
-	var seconds: Int
-		set(value) {
-			builderCopy.seconds = value
-		}
+    
+    var seconds: Long
+        set(value) {
+            builderCopy.seconds = value
+        }
 
-		get() = builderCopy.seconds
-	
-	
-	var nanos: Int
-		set(value) {
-			builderCopy.nanos = value
-		}
+        get() = builderCopy.seconds
+    
+    
+    var nanos: Int
+        set(value) {
+            builderCopy.nanos = value
+        }
 
-		get() = builderCopy.nanos
-	
-	
+        get() = builderCopy.nanos
+    
+    
 }
