+++ lib/src/commonMain/kotlin/com/google/protobuf/timestamp.builder.kt
@@ -4,16 +4,16 @@ import dev.mpr.grpc.ProtoDsl
 
 @ProtoDsl
 class TimestampBuilder {
-	constructor()
+    constructor()
 
-	constructor(copy: Timestamp) {
+    constructor(copy: Timestamp) {
 	    builderCopy.seconds = copy.seconds
 	    builderCopy.nanos = copy.nanos
 	    builderCopy.unknownFields = copy.unknownFields
 	}
 
 	private object builderCopy {
-	    var seconds: Int = 0
+	    var seconds: Long = 0L
 	    var nanos: Int = 0
 	    var unknownFields: ByteArray? = null
 	}
@@ -24,20 +24,21 @@ class TimestampBuilder {
     	builderCopy.unknownFields
 	)
 	
-	var seconds: Int
-		set(value) {
-			builderCopy.seconds = value
-		}
+    
+    var seconds: Long
+        set(value) {
+            builderCopy.seconds = value
+        }
 
-		get() = builderCopy.seconds
-	
-	
-	var nanos: Int
-		set(value) {
-			builderCopy.nanos = value
-		}
+        get() = builderCopy.seconds
+    
+    
+    var nanos: Int
+        set(value) {
+            builderCopy.nanos = value
+        }
 
-		get() = builderCopy.nanos
-	
-	
+        get() = builderCopy.nanos
+    
+    
 }
