+++ lib/src/commonMain/kotlin/com/google/protobuf/timestamp.data.kt
@@ -2,17 +2,17 @@ package com.google.protobuf
 
 
 data class Timestamp(
-    val seconds: Int = 0,
+    val seconds: Long = 0L,
     val nanos: Int = 0,
     val unknownFields: ByteArray? = null
 ) {
     companion object {
-		fun build(builder: TimestampBuilder.() -> Unit): Timestamp {
-			return TimestampBuilder().apply(builder).build()
-		}
+        fun build(builder: TimestampBuilder.() -> Unit): Timestamp {
+            return TimestampBuilder().apply(builder).build()
+        }
     }
 
-	fun copyBuild(builder: TimestampBuilder.() -> Unit): Timestamp {
-		return TimestampBuilder(this).apply(builder).build()
-	}
+    fun copyBuild(builder: TimestampBuilder.() -> Unit): Timestamp {
+        return TimestampBuilder(this).apply(builder).build()
+    }
 }
